refactor(tests): split test_error_helpers.py to meet LOC limit
Split exception-related tests into separate file:
- test_error_helpers.py: 294 lines (was 413)
- test_error_helpers_exception.py: 129 lines (new)

All 13 tests pass after split.
